Add the GNU GPL Version 3 License Documentation
This project has gone a long way since its first draft, and is now licensed under the GNU General Public License Version 3. This allows me to ensure that I retain full rights over this library. This is also a good hand to the Free Software developers out there, as it specifically disallows implementation of this library in proprietary software.