Allow NoValue to be hashable
While there's no real advantage to having NoValue hashable, it doesn't
make sense not to be, as it's a singleton and is not mutable. This can
also be used to have NoValue as part of a set. Since the class can only
ever have one instance, it's perfectly fine to pass in the class' hash.
It would only make little sense should the class be passed instead of
the instance, at which point the other comparison operators would fail
(yet hash() would still be equal). At that point it's up to the user to
use the singleton properly.