Implement hash checking for the Bypassers
It's not possible (or at the very least, not worth it) to look up the
items based solely on their hash values. Re-implementing the hash
functions wouldn't be worth the effort, so instead, simple lookup of
hashable items as well as comparing (and raising an exception if not
found) the different values can be used for a slower lookup method.