Add PEP 487 support
This adds support for __init_subclass__ and __set_name__, which we use
mostly to remove some boilerplate in the decorators, as well as the
stable view objects. This could also be used for the regular view
objects, but they are currently set after the class has been created,
rendering __set_name__ pretty useless for those.