Allow more dynamic creation of set instances
With _dict now being a class attribute, the relevant instance creation
method (__new__ if immutable, __init__ if mutable) of each set can use
it to define their mapping. Assigning the mapping to _dict when the
instance is created/initialized means the class attribute is effectively
shadowed (which is what we want). This allows subclasses to merely
define their prefered mapping at the class level, and the rest is done
automatically!